+++ src/titiler/mosaic/titiler/mosaic/extensions/__init__.py
@@ -1,0 +1,1 @@
+"""titiler mosaic extensions module."""